\name{Adressen}
\docType{data}
\alias{Adressen}
\title{Address data from the city of Dortmund}
\description{
  This data set gives all the addresses in the city of Dortmund. 
}
\usage{Adressen}
\format{A data.frame
  \tabular{lll}{
    STRNAME   \tab character \tab street name\cr
    STRSL \tab numeric \tab street number\cr
    HNR    \tab numeric \tab house number\cr
    HNRZ    \tab character \tab additional letter\cr
    RW   \tab numeric \tab longitude \cr
    HW     \tab numeric \tab latitude \cr
    UBZ     \tab numeric \tab subdistrict number
  }
}
\source{\url{https://open-data.dortmund.de}}
\keyword{datasets}