\name{df1}
\docType{data}
\alias{df1}
\title{Example dataset 1}
\description{
  This dataset contains separate street and house number information.
}
\usage{df1}
\format{A data.frame
  \tabular{lll}{
    gross_strasse   \tab character \tab street names\cr
    hausnr \tab character \tab house number and additional letter\cr
    Var1    \tab numeric \tab Variable 1\cr
    Var2    \tab character \tab Variable 2
  }
}
\source{Dortmunder Statistik}
\keyword{datasets}