\name{df2}
\docType{data}
\alias{df2}
\title{Example dataset 2}
\description{
  This dataset contains concatenated street and house number information.
}
\usage{df2}
\format{A data.frame
  \tabular{lll}{
    Adresse   \tab character \tab street name, house number and addional letter\cr
    Var1    \tab numeric \tab Variable 1\cr
    Var2    \tab character \tab Variable 2
  }
}
\source{Dortmunder Statistik}
\keyword{datasets}