\name{KOR.addrlink-package}
\alias{KOR.addrlink-package}
\alias{KOR.addrlink}
\docType{package}
\title{KOR.addrlink}
\description{Geocode address data from German municipalities}
\details{
\itemize{
  \item \code{\link{split_address}} Splits strings into street, house number and addional letter
  \item \code{\link{split_number}} Splits strings into house number and addional letter
  \item \code{\link{addrlink}} Matches splitted address data to reference table
}
Matching is based on street name, house number and additional letter. 
}
\author{Daniel Schürmann}
